move from hx711 to nau7802 + emc2302
rid ourselves of the direct management of fans,
saving four pins. get rid of the hx711. we're
going all-I2C, with the NAU7802 and EMC2302 for
our scale and fans, respectively. restore creation
of an I2C bus master. require version 0.0.3 of my
NAU7802 driver. commit to MOSFET for the motor.